Move request decoders to interop
